Assassin Auction Image Editing in Google Colab
Images were given light and dark outlines in Google Colab, coded in Python, using a Laplacian filter. The purpose of this process was to provide realistic images with a more pixelated look so they fit better with The Assassin Auction's 3D-retro art style. The greyscale image was first found by detecting how many pixels were a part of the red, green and blue channels and then averaging the values (dividing by the 3 channels). The greyscale image was then used to create an enhanced image using the Laplacian filter from the cv2 library.